backend/requirements.txt

amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
billiard==4.2.0
boto3==1.34.151
botocore==1.34.151
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.6.1
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.8
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
flower==2.0.1
humanize==4.10.0
prometheus_client==0.20.0
pytz==2024.1
tornado==6.4.1

argcomplete==1.10.3
beautifulsoup4==4.8.2
chardet==3.0.4
compressed-rtf==1.0.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
IMAPClient==2.1.0
lxml==5.2.2
olefile==0.47
pdfminer.six==20191110
pillow==10.4.0
pycryptodome==3.20.0
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
textract==1.6.5
typing_extensions==4.12.2
tzlocal==5.2
xlrd==1.2.0
XlsxWriter==3.2.0
